Fix bug on HibernateSessionExtension
Fix bug causing difficulty in using nested tests because of how
the extension was injecting a session.
Was previously using beforeEach to inject the session. But this
does not provide the testInstance of enclosing classes in nested
classes. This makes it (maybe) impossible to inject the session
into the enclosing classes when the context is of a nested class.