Fix bug in HibernateSessionExtension
Fix bug causing a hibernate session factory and session to be
created twice.
This was happening for nested tests. It was happening because
the initialization happens in junit5 extensions'
postProcessTestInstance method. For nested tests, this method
is called multiple times, depending on the degree of nesting. Thus
the hibernate session factory and session were being unitialized
twice, and that was causing test to fail is some situations with
an exception stating 'timeout while trying to lock table'.