Fix bug causing database entries not to be saved after first entry is saved.
The dependency injection configuration was such that each request
would be handled by a singleton endpoint. The endpoint is injected
with a service object which is injected with a repository object
which is injected with a hibernate session. It is the hibernate
session that is of most significance here.
Since the endpoint object handling each request was a singleton,
itself, and all the injected objects mentioned above (including the
hibernate session) were retained and used for every request.
That is the reason for the bug: the same session is used for every
request.
Why is that a problem?
For each request, when the request is received, a new session is
created for it and when the request has been handled, that same
session is commited (and should be closed. I forgot to do that,
which is veeeery wrong). Therefore, for the first request, all is
good, its session's transaction is committed to the database.
But for all subsequent, all was not so good. Whereas, the
application is commiting their session's transactions, the first
requests session is the one that was being used to perform DB
operations.

The solution: configure the DI to provide a new endpoint object,
service object and repository object for each request.