Fix bug causing saving of Visit entity not to cascade to children Payment entities
Since Session.save(...) is being used to save entities use Hibernates
@Cascade annotation to add the SAVE_UPDATE CascadeType.
An easier alternative would have been to use JPA's
@OneToMany(..., cascade = [CascadeType.ALL]) since it includes the
the SAVE_UPDATE CascadeType. But this would include
CascadeType.REMOVE which would not work with the logic for
finishing an ongoing visit. This logic requires that the child
Payment entities of a Visit should not be deleted.